src/services/ReSys.ML/requirements.txt
# Server
fastapi
uvicorn[standard]
pydantic
python-multipart

# Machine Learning
torch
sentence-transformers
numpy

# OpenTelemetry
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc


